Add `Issue` model to represent issue tracking data
A new `Issue` class was introduced in the `MaintenancePortal.Models` namespace. This class includes the following required properties: `Gene`, `Depth`, `Order`, `Title`, `Description`, `CreatedAt`, and `CreatedByUserId`. The `CreatedByUserId` property is marked with the `[ForeignKey("User")]` attribute to establish a foreign key relationship with the `User` entity.

Additionally, a computed property `DisplayId` was added, which formats `Gene`, `Depth`, and `Order` into a string representation. This property is excluded from database mapping using the `[NotMapped]` attribute.

The `System.ComponentModel.DataAnnotations.Schema` namespace was imported to support the use of `[ForeignKey]` and `[NotMapped]` attributes.