Add TicketsController with sample data and views
Introduced a new `TicketsController` class to manage ticket-related actions.
Added `Index` action to display a list of hardcoded tickets and a `New`
action to render a view for creating new tickets. Included necessary
`using` directives and set up a constructor for dependency injection,
though the `_repo` field is not yet utilized.