Disable unique index on Label.Name property
The unique index constraint on the `Name` property of the `Label`
entity in the `LabelConfiguration` class has been commented out.
This change removes the enforcement of unique values and the
associated index, which may impact database performance and
data integrity. The `builder.HasIndex(l => l.Name).IsUnique();`
line is now disabled.