Add Logout button for authenticated users in navbar
Added a conditional block in `_Layout.cshtml` to display a "Logout" button in the navigation bar when the user is authenticated. The button is styled as a `nav-link` and is aligned to the right using a `<ul>` with the class `navbar-nav ms-auto`. The logout functionality is implemented using a `<form>` that posts to the `Logout` action in the `User` controller.